Change battery data type to uint16_t from int. (#36)
Since battery data type will never result in negative values under proper operating conditions, I changed the data type from signed integer to unsigned integer.

+++ src/gen_cranberry/gc_board.cpp
@@ -142,7 +142,7 @@ static void gc_board_post(){
 
     // Check batt
     Serial.println(F("[P] Check batt value"));
-    int batt_val = gc_dev_batt_read();
+    uint16_t batt_val = gc_dev_batt_read();
 
     Serial.print(F("[P] batt value: "));
     Serial.print(batt_val);
