+++ src/gen_cranberry/gc_dev_batt.h
@@ -5,6 +5,6 @@
 #ifndef GC_DEV_BATT_H
 #define GC_DEV_BATT_H
 void gc_dev_batt_open(void);
-int gc_dev_batt_read_raw(void);
-int gc_dev_batt_read(void);
+uint16_t gc_dev_batt_read_raw(void);
+uint16_t gc_dev_batt_read(void);
 #endif
